+++ bukkit/src/main/java/de/maxhenkel/voicechat/plugins/impl/events/VoiceHostEventImpl.java
@@ -1,12 +1,17 @@
 package de.maxhenkel.voicechat.plugins.impl.events;
 
+import de.maxhenkel.voicechat.api.ServerPlayer;
 import de.maxhenkel.voicechat.api.events.VoiceHostEvent;
+import de.maxhenkel.voicechat.plugins.impl.ServerPlayerImpl;
+import org.bukkit.entity.Player;
 
 public class VoiceHostEventImpl extends ServerEventImpl implements VoiceHostEvent {
 
+    private final ServerPlayerImpl player;
     private String voiceHost;
 
-    public VoiceHostEventImpl(String voiceHost) {
+    public VoiceHostEventImpl(Player player, String voiceHost) {
+        this.player = new ServerPlayerImpl(player);
         this.voiceHost = voiceHost;
     }
 
@@ -19,4 +24,10 @@ public class VoiceHostEventImpl extends ServerEventImpl implements VoiceHostEven
     public void setVoiceHost(String voiceHost) {
         this.voiceHost = voiceHost;
     }
+
+    @Override
+    public ServerPlayer getPlayer() {
+        return player;
+    }
+
 }